Run RIPL in a separate context and dup the RubyTerminal cache so we don’t have a race condition.
Asynchronous callabcks store the `line' parameter they get for the
:enter event. This parameter however was a simple reference to the
RubyTerminal instance’s internal @cache, which after the callback
returned (asynchronous callbacks immediately return!) was cleared,
leaving the asynchronous callback with a reference to an empty string,
causing weird results and many nil outprints. Now the `line' parameter
is a proper dup-licate of the @cache.