Introduce EventedStorage, used for App#state.
This is basically a hash, but mixes in the Eventable module so
that global state changes are far more easy to listen to than
before (which required access to the target object, e.g. the map
widget needs access to the map window, which is bad). The new
functionality isn’t used yet, but will be soon.