Cannot change the name of a template after creation.
It is used to compare templates, and renaming a template would
effectively mean deleting an existing template and creating
a new one, because the name is the only thing used in Template#==
which in turn is used in some places to find specific templates
(e.g. the template combobox uses this).